refactor(Piece): 경로 찾는 기능을 템플릿 메소드를 사용하도록 변경
