+++ src/main/java/chess/domain/piece/Pawn.java
@@ -2,7 +2,6 @@ package chess.domain.piece;
 
 import chess.domain.Position;
 import chess.domain.Team;
-import java.util.ArrayList;
 import java.util.List;
 
 public class Pawn extends Piece {
@@ -46,36 +45,19 @@ public class Pawn extends Piece {
 
     @Override
     public List<Position> findPath(Position source, Position target) {
-        int rankDiff = source.rankDiff(target);
-        int fileDiff = source.fileDiff(target);
-
-        int count = initCount(rankDiff, fileDiff);
-        int rankUnit = rankDiff / count;
-        int fileUnit = fileDiff / count;
-
-        return makePath(count, rankUnit, fileUnit, source);
+        return findPathTemplate(source, target, this::calculateCount);
     }
 
-    private int initCount(int rankDiff, int fileDiff) {
-        int count = 1;
+    @Override
+    protected int calculateCount(int rankDiff, int fileDiff) {
         if (Math.abs(fileDiff) == Math.abs(rankDiff)) {
-            count = Math.abs(fileDiff);
+            return Math.abs(fileDiff);
         }
-        if (fileDiff == 0 || rankDiff == 0) {
-            count = Math.abs(fileDiff + rankDiff);
-        }
-
-        return count;
+        return Math.abs(fileDiff + rankDiff);
     }
 
-    private List<Position> makePath(int count, int rankUnit, int fileUnit, Position current) {
-        List<Position> path = new ArrayList<>();
-
-        for(int i = 0; i < count; i++) {
-            current = current.moveBy(rankUnit, fileUnit);
-            path.add(current);
-        }
-
-        return path;
+    @Override
+    public boolean isPawn() {
+        return true;
     }
 }
